Install a custom python3 kernel
This is so we can add the project directory to `sys.path`
consistently (which we do via the standard `PYTHONPATH` environment
variable).

This workaround is necessary because the `nbval` plugin for `py.test`
uses the same kernel as the notebook. Previously, we started the
notebook with `PYTHONPATH` in the environment; however, setting this
variable is not enough when using the `nbval` plugin, because that
does not pass on environment variables to the iPython kernel that it
starts.

This way we can be sure that anything using the `python3` kernel
within the Dockerised environment has the same `sys.path` behaviour.