Replace DEVELOPER instructions
This is a bit of shame because I'm removing a lot of obviously carefully
thought out content. However a lot of it no longer applies, both as a
result of the changes I've made and because the world generally shifting
underneath us. And I don't currently have the time to work through this
all, pick out the still relevant stuff and make all the necessary
changes.

Obviously the content is all still here in git so at some point we can
revisit this and see what can be salvaged.